pyproject.toml
[tool.poetry]
name = "telegram-bot-usn"
version = "0.1.0"
description = "Telegram bot for USN competition registration"
authors = ["alexkostromin <alexkostromin9797@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.4.1"
sqlalchemy = "^2.0.0"
aiosqlite = "^0.19.0"
python-dotenv = "^1.0.0"
django = "^4.2.8"
django-extensions = "^3.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
mypy = "^1.5.0"
black = "^23.9.0"
flake8 = "^6.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100
python_version = "310"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
